// mordisco-front/src/app/shared/models/notificacion/tipo-notificacion.ts
export enum TipoNotificacion {
  NUEVO_PEDIDO = 'NUEVO_PEDIDO',
  PEDIDO_EN_PREPARACION = 'PEDIDO_EN_PREPARACION',
  PEDIDO_LISTO_PARA_ENTREGAR = 'PEDIDO_LISTO_PARA_ENTREGAR',
  PEDIDO_LISTO_PARA_RETIRAR = 'PEDIDO_LISTO_PARA_RETIRAR',
  PEDIDO_EN_CAMINO = 'PEDIDO_EN_CAMINO',
  PEDIDO_COMPLETADO = 'PEDIDO_COMPLETADO',
  PEDIDO_CANCELADO = 'PEDIDO_CANCELADO',
  PAGO_CONFIRMADO = 'PAGO_CONFIRMADO',
  PAGO_RECHAZADO = 'PAGO_RECHAZADO'
}